tasks/fixture_downloader: move to spec/support, route GH through Octokit
Two related architectural changes in one PR.

== Move lib/fontisan/tasks -> spec/support

fontisan-the-gem never downloads fonts at runtime. The FixtureDownloader
was only ever invoked by `rake fixtures:download` (the test-suite setup
task). Shipping it under lib/ meant:

  - Test infrastructure appeared in the published gem's namespace.
  - Any future dev-only dependency it picked up would have to be
    rationalized as "not a runtime dep" - an awkward conversation.
  - The Rakefile's `require "fontisan/tasks"` worked from the source
    tree but would fail for an installed gem, hiding the fact that
    the task is project-internal.

Moved to spec/support/fixture_downloader.rb (alongside the existing
fixture_fonts.rb sibling). Renamed the class from
Fontisan::Tasks::FixtureDownloader to FixtureFonts::Downloader to
match the existing FixtureFonts module. Deleted:

  - lib/fontisan/tasks.rb
  - lib/fontisan/tasks/fixture_downloader.rb
  - lib/fontisan.rb `autoload :Tasks` line

The gemspec already excludes spec/ from the gem package, so the
published gem shrinks and gains zero new files.

== Route GitHub URLs through Octokit when GITHUB_TOKEN is present

The previous open-uri path's Authorization header gets stripped on
github.com -> raw.githubusercontent.com redirects (standard browser
security behavior in open-uri). The token never reaches the host
actually serving the bytes, so the rate-limit fix from PR #109
helped (it added retry) but didn't lift the 60/hr ceiling for
real. The Windows CI matrix still hit 429.

Octokit solves this properly:

  - Authenticates every request via its own HTTP machinery
    (Faraday middleware), independent of open-uri's redirect
    handling.
  - Auto-detects rate-limit responses and either retries or raises
    Octokit::RateLimitExceeded with reset info (no manual
    Retry-After parsing needed for GitHub URLs).
  - Addresses the right endpoint per URL shape - release asset API
    for /releases/download/... URLs, contents API for /raw/... URLs
    (with raw.githubusercontent.com fallback for files >1MB),
    authenticated GET for everything else.

Per the project rule "dev deps belong only in Gemfile, not gemspec":
Octokit is added to the Gemfile (group: :test), NOT to
fontisan.gemspec. Production installs never pull it in.

The downloader's routing logic:

  - If URL host is in GITHUB_HOSTS AND ENV["GITHUB_TOKEN"] is set
    AND `require "octokit"` succeeds -> use OctokitFetcher.
  - Otherwise -> fall back to open-uri (with Authorization header
    scoped to GitHub hosts via send_auth? guard).

The OctokitFetcher module dispatches by URL pattern:

  - github.com/owner/repo/raw/<ref>/<path>
      -> Octokit.contents (Base64-decoded) for files <=1MB,
         raw.githubusercontent.com fallback for larger files.
  - github.com/owner/repo/releases/download/<tag>/<asset>
      -> client.release_by_tag + release_asset API.
  - raw.githubusercontent.com, codeload.github.com, api.github.com,
    objects.githubusercontent.com -> authenticated GET.

Specs: 19 examples (was 17). New coverage:
  - "routes GitHub URLs through OctokitFetcher when token + Octokit
    present"
  - "falls back to open-uri when Octokit gem is not installed"
  - "does not route non-GitHub URLs through Octokit even when
    available"

All existing open-uri retry / Retry-After / credential-leak-guard
tests still pass.

== Why both changes together

Moving first would have left a working downloader with the same
redirect-stripping limitation. Adding Octokit first would have left
test infrastructure in lib/. Doing them together produces one clean
PR with the right final shape.
